ref(org): refactored org pkg to accomodate changes on internal and common pkgs [#10]
